+++ test/decode_test.dart
@@ -4,16 +4,29 @@ import 'package:test/test.dart';
 void main() {
   PostgreSQLConnection connection;
   setUp(() async {
-    connection = new PostgreSQLConnection("localhost", 5432, "dart_test",
-        username: "dart", password: "dart");
+    connection = new PostgreSQLConnection("localhost", 5432, "dart_test", username: "dart", password: "dart");
     await connection.open();
 
-    await connection.execute(
-        "CREATE TEMPORARY TABLE t (i int, s serial, bi bigint, bs bigserial, bl boolean, si smallint, t text, f real, d double precision, dt date, ts timestamp, tsz timestamptz)");
-    await connection.execute(
-        "INSERT INTO t (i, bi, bl, si, t, f, d, dt, ts, tsz) VALUES (-2147483648, -9223372036854775808, TRUE, -32768, 'string', 10.0, 10.0, '1983-11-06', '1983-11-06 06:00:00.000000', '1983-11-06 06:00:00.000000')");
-    await connection.execute(
-        "INSERT INTO t (i, bi, bl, si, t, f, d, dt, ts, tsz) VALUES (2147483647, 9223372036854775807, FALSE, 32767, 'a significantly longer string to the point where i doubt this actually matters', 10.25, 10.125, '2183-11-06', '2183-11-06 00:00:00.111111', '2183-11-06 00:00:00.999999')");
+    await connection.execute("""
+        CREATE TEMPORARY TABLE t (
+          i int, s serial, bi bigint, bs bigserial, bl boolean, si smallint, 
+          t text, f real, d double precision, dt date, ts timestamp, tsz timestamptz, j jsonb)
+    """);
+
+    await connection.execute("INSERT INTO t (i, bi, bl, si, t, f, d, dt, ts, tsz, j) "
+        "VALUES (-2147483648, -9223372036854775808, TRUE, -32768, "
+        "'string', 10.0, 10.0, '1983-11-06', "
+        "'1983-11-06 06:00:00.000000', '1983-11-06 06:00:00.000000', "
+        "'{\"key\":\"value\"}')");
+    await connection.execute("INSERT INTO t (i, bi, bl, si, t, f, d, dt, ts, tsz, j) "
+        "VALUES (2147483647, 9223372036854775807, FALSE, 32767, "
+        "'a significantly longer string to the point where i doubt this actually matters', "
+        "10.25, 10.125, '2183-11-06', '2183-11-06 00:00:00.111111', "
+        "'2183-11-06 00:00:00.999999', "
+        "'[{\"key\":1}]')");
+
+    await connection.execute("INSERT INTO t (i, bi, bl, si, t, f, d, dt, ts, tsz, j) "
+        "VALUES (null, null, null, null, null, null, null, null, null, null, null)");
   });
   tearDown(() async {
     await connection?.close();
@@ -24,6 +37,9 @@ void main() {
 
     var row1 = res[0];
     var row2 = res[1];
+    var row3 = res[2];
+
+    // lower bound row
     expect(row1[0], equals(-2147483648));
     expect(row1[1], equals(1));
     expect(row1[2], equals(-9223372036854775808));
@@ -38,17 +54,16 @@ void main() {
     expect(row1[9], equals(new DateTime.utc(1983, 11, 6)));
     expect(row1[10], equals(new DateTime.utc(1983, 11, 6, 6)));
     expect(row1[11], equals(new DateTime.utc(1983, 11, 6, 6)));
+    expect(row1[12], equals({"key": "value"}));
 
+    // upper bound row
     expect(row2[0], equals(2147483647));
     expect(row2[1], equals(2));
     expect(row2[2], equals(9223372036854775807));
     expect(row2[3], equals(2));
     expect(row2[4], equals(false));
     expect(row2[5], equals(32767));
-    expect(
-        row2[6],
-        equals(
-            "a significantly longer string to the point where i doubt this actually matters"));
+    expect(row2[6], equals("a significantly longer string to the point where i doubt this actually matters"));
     expect(row2[7] is double, true);
     expect(row2[7], equals(10.25));
     expect(row2[8] is double, true);
@@ -56,13 +71,32 @@ void main() {
     expect(row2[9], equals(new DateTime.utc(2183, 11, 6)));
     expect(row2[10], equals(new DateTime.utc(2183, 11, 6, 0, 0, 0, 111, 111)));
     expect(row2[11], equals(new DateTime.utc(2183, 11, 6, 0, 0, 0, 999, 999)));
+    expect(
+        row2[12],
+        equals([
+          {"key": 1}
+        ]));
+
+    // all null row
+    expect(row3[0], isNull);
+    expect(row3[1], equals(3));
+    expect(row3[2], isNull);
+    expect(row3[3], equals(3));
+    expect(row3[4], isNull);
+    expect(row3[5], isNull);
+    expect(row3[6], isNull);
+    expect(row3[7], isNull);
+    expect(row3[8], isNull);
+    expect(row3[9], isNull);
+    expect(row3[10], isNull);
+    expect(row3[11], isNull);
+    expect(row3[12], isNull);
   });
 
   test("Fetch/insert empty string", () async {
     await connection.execute("CREATE TEMPORARY TABLE u (t text)");
-    var results = await connection.query(
-        "INSERT INTO u (t) VALUES (@t:text) returning t",
-        substitutionValues: {"t": ""});
+    var results =
+        await connection.query("INSERT INTO u (t) VALUES (@t:text) returning t", substitutionValues: {"t": ""});
     expect(results, [
       [""]
     ]);
@@ -75,9 +109,8 @@ void main() {
 
   test("Fetch/insert null value", () async {
     await connection.execute("CREATE TEMPORARY TABLE u (t text)");
-    var results = await connection.query(
-        "INSERT INTO u (t) VALUES (@t:text) returning t",
-        substitutionValues: {"t": null});
+    var results =
+        await connection.query("INSERT INTO u (t) VALUES (@t:text) returning t", substitutionValues: {"t": null});
     expect(results, [
       [null]
     ]);